Further navigation improvements
