+++ tools/lib_alignments/__init__.py
@@ -1,2 +1,3 @@
 from tools.lib_alignments.media import AlignmentData, DetectedFace, Faces, Frames
-from tools.lib_alignments.jobs import Check, Draw, Extract, Manual, Reformat, RemoveAlignments
+from tools.lib_alignments.jobs import Check, Draw, Extract, Reformat, RemoveAlignments
+from tools.lib_alignments.jobs_manual import Manual
