+++ tools/lib_alignments/jobs.py
@@ -5,12 +5,10 @@ import os
 import pickle
 import struct
 from datetime import datetime
-import cv2
+from cv2 import circle, rectangle
 
-import numpy as np
 from tqdm import tqdm
 
-from lib.align_eyes import FACIAL_LANDMARKS_IDXS
 from plugins.PluginLoader import PluginLoader
 from . import DetectedFace, Faces, Frames
 
@@ -241,13 +239,13 @@ class Draw():
         top_left = (alignment["x"], alignment["y"])
         bottom_right = (alignment["x"] + alignment["w"],
                         alignment["y"] + alignment["h"])
-        cv2.rectangle(image, top_left, bottom_right, (0, 0, 255), 1)
+        rectangle(image, top_left, bottom_right, (0, 0, 255), 1)
 
     @staticmethod
     def draw_landmarks(image, landmarks):
         """ Draw the facial landmarks """
         for (pos_x, pos_y) in landmarks:
-            cv2.circle(image, (pos_x, pos_y), 1, (0, 255, 0), -1)
+            circle(image, (pos_x, pos_y), 1, (0, 255, 0), -1)
 
 
 class Extract():
@@ -323,362 +321,6 @@ class Extract():
             self.frames.save_image(self.faces_dir, output, resized_face)
             face_count += 1
         return face_count
-
-
-class Manual():
-    """ Manually adjust or create landmarks data """
-    def __init__(self, alignments, arguments):
-        self.verbose = arguments.verbose
-        self.alignments = alignments
-        self.faces = self.set_output(arguments.faces_dir)
-        self.frames = Frames(arguments.frames_dir, self.verbose)
-        self.align_eyes = arguments.align_eyes
-        self.extractor = PluginLoader.get_extractor("Align")()
-        self.state = self.set_state()
-        self.controls = self.set_controls()
-        self.colors = {1: (255, 0, 0),
-                       2: (0, 255, 0),
-                       3: (0, 0, 255),
-                       4: (255, 255, 0),
-                       5: (255, 0, 255),
-                       6: (0, 255, 255)}
-
-    def set_output(self, faces_folder):
-        """ Set the output to be an existing or new folder """
-        if not os.path.isdir(faces_folder):
-            print("Creating output folder at {}".format(faces_folder))
-            os.makedirs(faces_folder)
-        return Faces(faces_folder, self.verbose)
-
-    def set_state(self):
-        """ Set the initial display state """
-        state = {"bounding_box": dict(),
-                 "extract_box": dict(),
-                 "landmarks": dict(),
-                 "landmarks_mesh": dict(),
-                 "image": dict(),
-                 "select": None}
-
-        color = 1
-        for key in sorted(state.keys()):
-            if key == "select":
-                continue
-            state[key]["display"] = bool(key == "image")
-            if key == "image":
-                continue
-            state[key]["size"] = 1
-            state[key]["color"] = color
-            color = color + 1 if color != 6 else 1
-
-        return state
-
-    def set_controls(self):
-        """ Set keyboard controls, destination and help text """
-        controls = {ord("z"): {"action": "prev",
-                               "args": (-1, "navigation"),
-                               "help": "Previous Frame"},
-                    ord("x"): {"action": "next",
-                               "args": (1, "navigation"),
-                               "help": "Next Frame"},
-                    27: {"action": "quit",
-                         "args": (None, "navigation"),
-                         "help": "Exit"},
-                    ord("y"): {"action": self.toggle_state,
-                               "args": ("image", "display"),
-                               "help": ("Toggle Image")},
-                    ord("u"): {"action": self.toggle_state,
-                               "args": ("bounding_box", "display"),
-                               "help": ("Toggle Bounding Box")},
-                    ord("i"): {"action": self.toggle_state,
-                               "args": ("extract_box", "display"),
-                               "help": ("Toggle Extract Box")},
-                    ord("o"): {"action": self.toggle_state,
-                               "args": ("landmarks", "display"),
-                               "help": ("Toggle Landmarks")},
-                    ord("p"): {"action": self.toggle_state,
-                               "args": ("landmarks_mesh", "display"),
-                               "help": ("Toggle Landmarks Mesh")},
-                    ord("h"): {"action": self.iterate_state,
-                               "args": ("bounding_box", "color"),
-                               "help": ("Cycle Bounding Box Color")},
-                    ord("j"): {"action": self.iterate_state,
-                               "args": ("extract_box", "color"),
-                               "help": ("Cycle Extract Box Color")},
-                    ord("k"): {"action": self.iterate_state,
-                               "args": ("landmarks", "color"),
-                               "help": ("Cycle Landmarks Color")},
-                    ord("l"): {"action": self.iterate_state,
-                               "args": ("landmarks_mesh", "color"),
-                               "help": ("Cycle Landmarks Mesh Color")},
-                    ord("v"): {"action": self.iterate_state,
-                               "args": ("bounding_box", "size"),
-                               "help": ("Cycle Bounding Box thickness")},
-                    ord("b"): {"action": self.iterate_state,
-                               "args": ("extract_box", "size"),
-                               "help": ("Cycle Extract Box thickness")},
-                    ord("n"): {"action": self.iterate_state,
-                               "args": ("landmarks", "size"),
-                               "help": ("Cycle Landmarks point size")},
-                    ord("m"): {"action": self.iterate_state,
-                               "args": ("landmarks_mesh", "size"),
-                               "help": ("Cycle Landmarks Mesh thickness")},
-                    (ord("0"), ord("9")): {
-                        "action": self.set_state_value,
-                        "args": ("select", "select"),
-                        "help": "Select/Deselect face at this index"}}
-
-        return controls
-
-    def process(self):
-        """ Process manual extraction """
-        self.generate_help()
-        self.display_frames()
-
-    def generate_help(self):
-        """ Generate help output """
-        sections = ("navigation", "select", "display", "color", "size")
-        helpout = {section: list() for section in sections}
-        for key, val in self.controls.items():
-            if isinstance(key, tuple):
-                helpout[val["args"][1]].append(
-                    (val["help"],
-                     "{} to {}".format(chr(key[0]), chr(key[1]))))
-                continue
-            helpout[val["args"][1]].append((val["help"], chr(key)))
-
-        for section in sections:
-            spacer = "=" * int((40 - len(section)) / 2)
-            display = "{} {} {}\n".format(spacer, section.upper(), spacer)
-            helpsection = sorted(helpout[section])
-            if section == "navigation":
-                helpsection = sorted(helpout[section], reverse=True)
-            for item in helpsection:
-                key = item[1] if item[0] != "Exit" else "ESC"
-                display += "  - '{}': {}\n".format(key, item[0])
-
-            print(display)
-
-    def toggle_state(self, item, category):
-        """ Toggle state of requested item """
-        self.state[item][category] = not self.state[item][category]
-
-    def iterate_state(self, item, category):
-        """ Cycle through options (6 possible) """
-        max_val = 6 if category == "color" else 3
-        val = self.state[item][category]
-        val = val + 1 if val != max_val else 1
-        self.state[item][category] = val
-
-    def set_state_value(self, item, value):
-        """ Set state of requested item or toggle off """
-        state = self.state[item]
-        if state == value:
-            self.state[item] = None
-        else:
-            self.state[item] = value
-
-    def frame_selector(self, idx):
-        """ Return frame at given index """
-        frame = self.frames.file_list_sorted[idx]
-        fullpath = os.path.join(self.frames.folder, frame)
-        return frame, fullpath
-
-    def get_frame(self, idx, rois):
-        """ Compile the frame """
-        frame, fullpath = self.frame_selector(idx)
-        alignments = self.alignments.alignments.get(frame, list())
-        img = cv2.imread(fullpath)
-        if not self.state["image"]["display"]:
-            img = self.black_image(img)
-        if self.state["bounding_box"]["display"]:
-            self.draw_bounding_box(img, alignments)
-        if self.state["extract_box"]["display"]:
-            self.draw_extract_box(img, rois)
-        if self.state["landmarks"]["display"]:
-            self.draw_landmarks(img, alignments)
-        if self.state["landmarks_mesh"]["display"]:
-            self.draw_landmarks_mesh(img, alignments)
-        if (self.state["select"] and
-                int(self.state["select"]) < len(alignments)):
-            self.grey_out_faces(img, rois)
-        return frame, img, alignments
-
-    @staticmethod
-    def black_image(image):
-        """ Return black image to correct dimensions """
-        height, width = image.shape[:2]
-        image = np.zeros((height, width, 3), np.uint8)
-        return image
-
-    def draw_bounding_box(self, image, alignments):
-        """ Draw the bounding box around faces """
-        color = self.colors[self.state["bounding_box"]["color"]]
-        thickness = self.state["bounding_box"]["size"]
-        for alignment in alignments:
-            top_left = (alignment["x"], alignment["y"])
-            bottom_right = (alignment["x"] + alignment["w"],
-                            alignment["y"] + alignment["h"])
-            cv2.rectangle(image, top_left, bottom_right, color, thickness)
-
-    def draw_extract_box(self, image, rois):
-        """ Draw the extracted face box """
-        if not rois:
-            return
-        color = self.colors[self.state["extract_box"]["color"]]
-        thickness = self.state["extract_box"]["size"]
-
-        for roi in rois:
-            cv2.polylines(image, roi, True, color, thickness)
-
-    def draw_landmarks(self, image, alignments):
-        """ Draw the facial landmarks """
-        color = self.colors[self.state["landmarks"]["color"]]
-        radius = self.state["landmarks"]["size"]
-
-        for alignment in alignments:
-            landmarks = alignment["landmarksXY"]
-            for (pos_x, pos_y) in landmarks:
-                cv2.circle(image, (pos_x, pos_y), radius, color, -1)
-
-    def draw_landmarks_mesh(self, image, alignments):
-        """ Draw the facial landmarks """
-        color = self.colors[self.state["landmarks_mesh"]["color"]]
-        thickness = self.state["landmarks_mesh"]["size"]
-
-        for alignment in alignments:
-            landmarks = alignment["landmarksXY"]
-            for key, val in FACIAL_LANDMARKS_IDXS.items():
-                points = np.array([landmarks[val[0]:val[1]]], np.int32)
-                fill_poly = bool(key in ("right_eye", "left_eye", "mouth"))
-                cv2.polylines(image, points, fill_poly, color, thickness)
-
-    def grey_out_faces(self, image, rois):
-        """ Grey out all faces except target """
-        overlay = image.copy()
-        for idx, roi in enumerate(rois):
-            if idx != int(self.state["select"]):
-                cv2.fillPoly(overlay, roi, (0, 0, 0))
-        alpha = 0.6
-        cv2.addWeighted(overlay, alpha, image, 1 - alpha, 0, image)
-
-    def display_frames(self):
-        """ Iterate through frames """
-        idx = 0
-        max_idx = self.frames.count - 1
-        cv2.namedWindow("Frame")
-
-        matrices = list()
-        range_keys = dict()
-
-        frame, img, alignments = self.get_frame(idx, matrices)
-
-        for keyrange, value in self.controls.items():
-            if not isinstance(keyrange, tuple):
-                continue
-            for key in range(keyrange[0], keyrange[1] + 1):
-                range_keys[key] = value
-
-        while True:
-            if cv2.getWindowProperty('Frame', cv2.WND_PROP_VISIBLE) < 1:
-                break
-            rois = self.get_faces(img, alignments)
-            cv2.imshow("Frames", img)
-            key = cv2.waitKey(100)
-
-            if key in self.controls.keys():
-                action = self.controls[key]["action"]
-                args = self.controls[key]["args"]
-                if action == "quit":
-                    break
-                elif action in ("prev", "next"):
-                    end = 0 if action == "prev" else max_idx
-                    idx += args[0] if idx != end else 0
-                else:
-                    action(*args)
-            elif key in range_keys.keys():
-                action = range_keys[key]["action"]
-                args = range_keys[key]["args"][0]
-                action(args, chr(key))
-
-            frame, img, alignments = self.get_frame(idx, rois)
-
-        cv2.destroyAllWindows()
-
-    def get_faces(self, image, frame_alignments):
-        """ Display associated face """
-        size = 256  # Standard Faceswap size
-        faces = list()
-        rois = list()
-        total_alignments = len(frame_alignments)
-
-        for idx, alignment in enumerate(frame_alignments):
-            face = DetectedFace(image,
-                                alignment["r"],
-                                alignment["x"],
-                                alignment["w"],
-                                alignment["y"],
-                                alignment["h"],
-                                alignment["landmarksXY"])
-
-            resized_face, f_align = self.extractor.extract(image,
-                                                           face,
-                                                           size,
-                                                           self.align_eyes)
-            rois.append(self.original_roi(f_align, size))
-
-            if idx % 4 == 0:
-                row = resized_face
-            else:
-                row = np.concatenate((row, resized_face), axis=1)
-
-            if (idx + 1) % 4 == 0 or idx + 1 == total_alignments:
-                faces.append(row)
-
-        image = self.compile_faces_image(faces, total_alignments, size)
-        self.show_hide_faces(bool(faces), image)
-        return rois
-
-    @staticmethod
-    def original_roi(matrix, size):
-        """ Return the original ROI of an extracted face """
-        padding = 48    # Faceswap padding
-
-        points = np.array([[0, 0], [0, size - 1],
-                           [size - 1, size - 1], [size - 1, 0]],
-                          np.int32)
-        points = points.reshape((-1, 1, 2))
-
-        matrix = matrix * (size - 2 * padding)
-        matrix[:, 2] += padding
-        matrix = cv2.invertAffineTransform(matrix)
-
-        return [cv2.transform(points, matrix)]
-
-    @staticmethod
-    def compile_faces_image(faces, total_faces, size):
-        """ Compile the faces into tiled image """
-        image = None
-        blank_face = np.zeros((size, size, 3), np.uint8)
-        total_rows = len(faces)
-        remainder = 4 - (total_faces % 4)
-        for idx, row in enumerate(faces):
-            if idx + 1 == total_rows and remainder != 4:
-                for _ in range(remainder):
-                    row = np.concatenate((row, blank_face), axis=1)
-            if idx == 0:
-                image = row
-                continue
-            image = np.concatenate((image, row), axis=0)
-        return image
-
-    @staticmethod
-    def show_hide_faces(display, image):
-        """ Show or remove window if faces are available """
-        if display:
-            cv2.namedWindow("Faces")
-            cv2.imshow("Faces", image)
-        elif cv2.getWindowProperty("Faces", cv2.WND_PROP_VISIBLE) == 1.0:
-            cv2.destroyWindow("Faces")
 
 
 class Reformat():
